convert the level processor to standard es6 imports
